Async impl
This is a simplistic implementation that uses tokio
as a threadpool rather than a fully-fledged asynchronous
/concurrent task processing system.

Essentially, use a global mutex around a map to store
outstanding requests from each thread. When a thread
receives a packet (whether the originating thread from
the original request or not), check if the response
matches a currently outstanding request, and print the
response if required.

There's also a bit more bookkeeping within the task for
each host to attempt to track the outstanding requests
that have been sent. Outstanding requests are only
tracked in the thread they originate from to reduce
synchronization costs.

+ test cases
+ the readme